refactor(events): split mega lib module
